[javascript] Move rules and tests to new packages
