+++ pmd-javascript/src/main/java/net/sourceforge/pmd/lang/ecmascript/rule/bestpractices/ConsistentReturnRule.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.ecmascript.rule.basic;
+package net.sourceforge.pmd.lang.ecmascript.rule.bestpractices;
 
 import java.util.List;
