--- pmd-javascript/src/test/java/net/sourceforge/pmd/lang/ecmascript/rule/controversial/ControversialRulesTest.java
@@ -1,17 +1,0 @@
-/**
- * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
- */
-
-package net.sourceforge.pmd.lang.ecmascript.rule.controversial;
-
-import net.sourceforge.pmd.testframework.SimpleAggregatorTst;
-
-public class ControversialRulesTest extends SimpleAggregatorTst {
-
-    private static final String RULESET = "rulesets/ecmascript/controversial.xml";
-
-    @Override
-    public void setUp() {
-        addRule(RULESET, "AvoidWithStatement");
-    }
-}
