--- pmd-javascript/src/test/java/net/sourceforge/pmd/lang/ecmascript/rule/unnecessary/UnnecessaryRulesTest.java
@@ -1,19 +1,0 @@
-/**
- * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
- */
-
-package net.sourceforge.pmd.lang.ecmascript.rule.unnecessary;
-
-import net.sourceforge.pmd.testframework.SimpleAggregatorTst;
-
-public class UnnecessaryRulesTest extends SimpleAggregatorTst {
-
-    private static final String RULESET = "rulesets/ecmascript/unnecessary.xml";
-
-    @Override
-    public void setUp() {
-        addRule(RULESET, "UnnecessaryBlock");
-        addRule(RULESET, "UnnecessaryParentheses");
-        addRule(RULESET, "NoElseReturn");
-    }
-}
