+++ pmd-javascript/src/test/resources/net/sourceforge/pmd/lang/ecmascript/rule/bestpractices/xml/ScopeForInVariable.xml
