+++ pmd-javascript/src/test/resources/net/sourceforge/pmd/lang/ecmascript/rule/codestyle/xml/IfElseStmtsMustUseBraces.xml
